tesis_api/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
openai==1.6.1
tiktoken==0.5.2
numpy==1.26.2
tqdm==4.66.1
